% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_names}
\alias{clean_names}
\title{Clean a vector of names}
\usage{
clean_names(x, bad_names = c())
}
\arguments{
\item{x}{character vector}

\item{bad_names}{a character vector of names in X to turn to NA}
}
\description{
Clean a vector of names
}
\details{
Does not do splits
}
