% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{compute_variables}
\alias{compute_variables}
\title{Create variables for linkage}
\usage{
compute_variables(input, sex = F)
}
\arguments{
\item{input}{data.table of pairs from two datasets prepped with prep_data_for_linkage}

\item{sex}{logical. Should sex/gender based variables be computed?}
}
\description{
Create variables for linkage
}
