% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_two}
\alias{fake_two}
\title{Fake data for linkage testing}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9561 rows and 16 columns.
}
\usage{
fake_two
}
\description{
Fake data for linkage testing
}
\keyword{datasets}
