% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fillblanks}
\alias{fillblanks}
\title{Replace NA entries with the other valid entries}
\usage{
fillblanks(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Replace NA entries with the other valid entries
}
