% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_link.R
\name{fit_link}
\alias{fit_link}
\title{Fit a linkage model (screening model + a stacked ensemble}
\usage{
fit_link(train, formula, ensemble = default_ensemble())
}
\arguments{
\item{train}{data.frame containing the training data}

\item{formula}{formula. The formula to fit}

\item{ensemble}{a named list of parsnip model specifications}
}
\description{
Fit a linkage model (screening model + a stacked ensemble
}
