% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_starter.R
\name{generate_starter}
\alias{generate_starter}
\title{Generate an ensemble based on fake data to use as training dataset creation helper}
\usage{
generate_starter(
  variables = c("first_name", "last_name", "date_of_birth"),
  match_prop = 0.3
)
}
\arguments{
\item{variables}{Character. One or more of 'first_name', 'last_name', 'date_of_birth', 'sex', 'middle_initial', 'zip', 'phone', and 'location.'
Indicates whether these variables (as prepped by prep_data_for_linkage and compute_variables) should inform the resulting linkage ensemble}

\item{match_prop}{numeric (0,1). The proportion of the training dataset that is made up of matches}
}
\description{
Generate an ensemble based on fake data to use as training dataset creation helper
}
