% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_functions.R
\name{get_chrome_list}
\alias{get_chrome_list}
\title{Gets List of Relevant Chrome Versions}
\usage{
get_chrome_list()
}
\description{
This is a helper function to be used alongside RSelenium. Returns a list of
    available chrome versions on the user's computer
}
\details{
RSelenium often crashes when there is an incorrect chrome association, so this
    is a great helper function that allows us to more quickly and accurately guess chrome
    versions without too much pain.
}
