% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{is_error}
\alias{is_error}
\title{Logical Inherits Error}
\usage{
is_error(x)
}
\arguments{
\item{x}{An object to test for whether an error occured}
}
\description{
This function simply returns a logical of whether or not an object
    resulted in an error, whether it be from the use of [try] or [tryCatch].
}
\details{
There are many times when we want to determine if a try-error or a 
    standard error occured. This helps make conditionals more streamlined, 
    but needs to be paired with the functions that will create objects with
    an error class, such as [try] and [tryCatch]
}
