% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{len0}
\alias{len0}
\title{Logical Length 0}
\usage{
len0(x)
}
\arguments{
\item{x}{A one-dimensional object to test length}
}
\description{
This function simply returns a logical of whether or not the length
    of an object is equal to zero. Can be used for any single dimensional object.
}
\details{
Often the only reason we care about length of an object is to check 
   whether the object has any dimensions, as many times a length of 0 is more
   the result of either an error or lack of data. Returning a logical for length
   helps make conditionals much more streamlined.
}
