% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_history_two}
\alias{location_history_two}
\title{Location History two}
\format{
An object of class \code{sf} (inherits from \code{data.table}, \code{data.frame}) with 11949 rows and 3 columns.
}
\usage{
location_history_two
}
\description{
Location History two
}
\keyword{datasets}
