% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_hash}
\alias{make_hash}
\title{Make a row-wise hash from a data.frame (or list)}
\usage{
make_hash(x)
}
\arguments{
\item{x}{single level list (e.g. data.frame)}
}
\description{
Make a row-wise hash from a data.frame (or list)
}
