% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmaker.R
\name{matchmaker}
\alias{matchmaker}
\title{A shiny app to facilitate the creation of a manual labelling of matches}
\usage{
matchmaker(dir = NULL, ...)
}
\arguments{
\item{dir}{the directory to have the file selector start in}

\item{...}{passed to shiny::shinyApp}
}
\description{
A shiny app to facilitate the creation of a manual labelling of matches
}
