% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nb}
\alias{nb}
\title{Replace NA with blanks ('')}
\usage{
nb(x)
}
\arguments{
\item{x}{a character vector (or something to be converted to character)}
}
\description{
Replace NA with blanks ('')
}
\keyword{internal}
