% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hyrule_data.R
\name{new_hyrule_data}
\alias{new_hyrule_data}
\alias{new_hyrule_data.data.frame}
\alias{new_hyrule_data.character}
\alias{new_hyrule_data.Dataset}
\alias{new_hyrule_data.DBIconnection}
\title{Create a new hyrule data object}
\usage{
new_hyrule_data(src, ...)

\method{new_hyrule_data}{data.frame}(src, ...)

\method{new_hyrule_data}{character}(src, ...)

\method{new_hyrule_data}{Dataset}(src, ...)

\method{new_hyrule_data}{DBIconnection}(src, from, ...)
}
\arguments{
\item{src}{One of a data.frame type object,
vector of .parquet file paths, or a DBIConnection object}

\item{...}{arguments passed onto lower level functions for}

\item{table_name}{when `src` is a DBIConnection, a table name using I or DBI::Id. See dbplyr \href{https://dbplyr.tidyverse.org/reference/tbl.src_dbi.html}{documentation} for more info}
}
\description{
Create a new hyrule data object
}
\details{
stuff
}
