% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{nrow0}
\alias{nrow0}
\title{Logical Number of Row 0}
\usage{
nrow0(x)
}
\arguments{
\item{x}{A matrix or data.frame to test number of rows}
}
\description{
This function simply returns a logical of whether or not the number
    of rows of a data.frame or matrix is equal to zero. Must be a data.frame or
    matrix to make sense.
}
\details{
Often the only reason we care about the number of rows a data.frame has
    (or technically a matrix) is to see if there is actually any data in the object.
    If there is not any data, it doesn't make sense to continue processes generally,
    so returning a logical for nrow helps make conditionals more streamlined
}
