% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairs}
\alias{pairs}
\title{Example pairs of matches}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 125 rows and 3 columns.
}
\usage{
pairs
}
\description{
Example pairs of matches
}
\keyword{datasets}
