% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hyrule_link}
\alias{predict.hyrule_link}
\title{Predicting with a hyrule_link}
\usage{
\method{predict}{hyrule_link}(object, new_data, members = F, opts = list(), ...)
}
\arguments{
\item{object}{hyrule_link object}

\item{new_data}{data.frame data containing variables to predict on}

\item{members}{logical. Whether predictions from the ensemble members should be returned}

\item{opts}{list. OPtion arguments to be passed on to predict.model_stack which passes it on to the prediction routines of the underlying ensemble}

\item{...}{not implemented}
}
\description{
Predicting with a hyrule_link
}
\details{
predict.hyrule_link always requests "prob" type from predict.model_stack
}
