% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_linkage.R
\name{prep_data_for_linkage}
\alias{prep_data_for_linkage}
\title{Prepare a dataset for linkage}
\usage{
prep_data_for_linkage(
  d,
  first_name = NULL,
  last_name = NULL,
  dob = NULL,
  middle_name = NULL,
  zip = NULL,
  ssn = NULL,
  sex = NULL,
  name_heuristics = TRUE
)
}
\arguments{
\item{d}{data.frame}

\item{first_name}{character. Identifies the column in `d` with the first name}

\item{last_name}{character. Identifies the column in `d` with the last name}

\item{dob}{character. Identifies the column with date of birth}

\item{middle_name}{character identifying the column with middle name (or middle initial)}

\item{zip}{character. Default is NULL. Column that includes ZIP code data}

\item{ssn}{character: column of social security number}

\item{sex}{character: Column specifiyng}

\item{name_heuristics}{logical. Indicates whether common name cleaning heuristics should be implemented}
}
\description{
Prepare a dataset for linkage
}
\details{
This function cleans/formats commonly used variables (depending on if they were supplied or not).
}
