% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_location_history.R
\name{process_location_history}
\alias{process_location_history}
\title{Compute address/location related metrics for record linkage}
\usage{
process_location_history(pairs, xy1, id1, xy2, id2)
}
\arguments{
\item{pairs}{data.frame/data.table of record pairs to generate variables for. Must include two id columns (specified by `id1` and `id2` arguments) for `d1` and `d2`}

\item{xy1}{sf. Coordinates per record for `d1` for those rows that have them. Must be in the same CRS as `xy2` and have an `id1` column.}

\item{id1}{character. Column name identifying a ID column in `id1`. Also must exist in `pairs` and `xy1` and `xy1`, if using.}

\item{xy2}{sf. Coordinates per record for  `d2` for those rows that have them. Must be in the same CRS as `xy1` and have an `id2` column.}

\item{id2}{character. Column name identifying a ID column in `id2`. Also must exist in `pairs` and `xy2` and `xy2`, if using.}
}
\description{
Compute address/location related metrics for record linkage
}
\details{
Given record pairs and location histories, this function returns a data.table(id1, id2, exact_location, min_loc_distance).
`exact_location` is a binary flag indicating that the pair has at least one geocoded location within 3 meters.
`min_loc_distance` is the minimum distance observed between a record pair in meters
}
