% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{trans_data_frame}
\alias{trans_data_frame}
\title{Transform Data Frame}
\usage{
trans_data_frame(x, strings_as_factors = F)
}
\arguments{
\item{x}{A data.frame object, typically.}

\item{strings_as_factors}{Logical value for making strings factors. Default to FALSE.}
}
\description{
Similar to the standard [t] function except that it preserves the data.frame.
    [t] has the nasty consequence of transforming a data.frame into a matrix, which is 
    often not desired.
}
\details{
One of the most important abilities with data frames is to swap its dimensions.
    However, when the dimensions are swapped via [t], the object is always switched to a
    matrix, which isn't always desirable. This function simplifies the process by transforming,
    forcing into a data.frame, and then removes any uselsess row names that would otherwise
    be associated with the process
}
