% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_names}
\alias{clean_names}
\title{Clean a vector of names}
\usage{
clean_names(x, type = "first")
}
\arguments{
\item{x}{character vector}

\item{type}{one of 'first' or last}
}
\description{
Clean a vector of names
}
\details{
Does not do splits
}
