% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_one}
\alias{fake_one}
\title{Fake data for linkage testing}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5511 rows and 13 columns.
}
\usage{
fake_one
}
\description{
Fake data for linkage testing
}
\keyword{datasets}
