% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_link.R
\name{fit_link}
\alias{fit_link}
\title{Fit a linkage model (screening model + a stacked ensemble}
\usage{
fit_link(
  train,
  formula,
  ensemble = default_ensemble(),
  bounds = c(0.001, 0.999),
  stage_one_only = FALSE
)
}
\arguments{
\item{train}{data.frame containing the training data}

\item{formula}{formula. The formula to fit}

\item{ensemble}{a named list of parsnip model specifications}

\item{bounds}{numeric. Two numbers (0-1) representing the predicted probability bounds from the screening model that should be fit with the ensemble.
Results outside these bounds are considered clear enough (not-)matches to not need further processing.}

\item{stage_one_only}{logical. Whether to return the stage one results so that models can be custom implemented with the larger tidy models universe}
}
\description{
Fit a linkage model (screening model + a stacked ensemble
}
